+++ src/Service/AlertManager.php
@@ -2,15 +2,10 @@
 
 namespace App\Service;
 
-use App\Exception\MailException;
-use App\Exception\ParsingException;
 use App\Model\Application;
+use App\Model\Device;
 use App\Model\Gateway;
 use ItkDev\MetricsBundle\Service\MetricsService;
-use Symfony\Contracts\HttpClient\Exception\ClientExceptionInterface;
-use Symfony\Contracts\HttpClient\Exception\RedirectionExceptionInterface;
-use Symfony\Contracts\HttpClient\Exception\ServerExceptionInterface;
-use Symfony\Contracts\HttpClient\Exception\TransportExceptionInterface as HttpClientTransportExceptionInterface;
 
 final readonly class AlertManager
 {
@@ -19,6 +14,7 @@ final readonly class AlertManager
         private SmsClient $smsClient,
         private MailService $mailService,
         private MetricsService $metricsService,
+        private TemplateService $templateService,
         private bool $applicationCheckStartDate,
         private int $gatewayLimit,
         private string $gatewayFallbackMail,
@@ -34,24 +30,7 @@ final readonly class AlertManager
     ) {
     }
 
-    /**
-     * Check gateways.
-     *
-     * @param \DateTimeImmutable $now
-     *   The timestamp to use for "now" time
-     * @param bool $filterOnStatus
-     *   Indicates whether to filter gateways based on a specific status
-     *
-     * @throws ClientExceptionInterface
-     * @throws RedirectionExceptionInterface
-     * @throws ServerExceptionInterface
-     * @throws MailException
-     * @throws ParsingException
-     * @throws \DateInvalidTimeZoneException
-     * @throws \DateMalformedStringException
-     * @throws HttpClientTransportExceptionInterface
-     */
-    public function checkGateways(\DateTimeImmutable $now, bool $filterOnStatus = true): void
+    public function checkGateways(\DateTimeImmutable $now, bool $filterOnStatus = true, string $overrideMail = '', string $overridePhone = ''): void
     {
         $gateways = $this->apiClient->getGateways($filterOnStatus);
         foreach ($gateways as $gateway) {
@@ -64,7 +43,7 @@ final readonly class AlertManager
                 );
                 // Gateway limit for last seen is reached.
                 $this->mailService->sendEmail(
-                    to: $this->findGatewayToMailAddress($gateway),
+                    to: $this->findGatewayToMailAddress($gateway, $overrideMail),
                     context: [
                         'gateway' => $gateway,
                         'diff' => $diff,
@@ -79,12 +58,23 @@ final readonly class AlertManager
                     textTemplate: 'gateway.txt.twig',
                 );
 
-                // @todo: SMS?
+                // SMS
+                $this->smsClient->send(
+                    to: [$this->findGatewayPhone($gateway, $overridePhone)],
+                    message: $this->templateService->renderTemplate('sms/gateway.twig', [
+                        'gateway' => $gateway,
+                        'since' => [
+                            'hours' => floor($diff / 3600),
+                            'minutes' => floor(($diff % 3600) / 60),
+                        ],
+                        'url' => $this->gatewayBaseUrl.$gateway->gatewayId,
+                    ])
+                );
             }
         }
     }
 
-    public function checkApplications(\DateTimeImmutable $now, bool $filterOnStatus = true): void
+    public function checkApplications(\DateTimeImmutable $now, bool $filterOnStatus = true, string $overrideMail = '', string $overridePhone = ''): void
     {
         $apps = $this->apiClient->getApplications($filterOnStatus);
         foreach ($apps as $app) {
@@ -93,12 +83,12 @@ final readonly class AlertManager
             }
 
             foreach ($app->devices as $deviceId) {
-                $this->checkDevice($now, $deviceId);
+                $this->checkDevice($now, $deviceId, $app, $overrideMail, $overridePhone);
             }
         }
     }
 
-    public function checkDevice(\DateTimeImmutable $now, int $deviceId, ?Application $application = null): void
+    public function checkDevice(\DateTimeImmutable $now, int $deviceId, ?Application $application = null, string $overrideMail = '', string $overridePhone = ''): void
     {
         $device = $this->apiClient->getDevice($deviceId);
 
@@ -108,8 +98,9 @@ final readonly class AlertManager
             $this->metricsService->counter(
                 name: 'alter_message_missing_total',
                 help: 'Device is missing latest received message',
-                labels: [ 'type' => 'info', 'id' => $device->id]
+                labels: ['type' => 'info', 'id' => $device->id]
             );
+
             return;
         }
 
@@ -117,9 +108,6 @@ final readonly class AlertManager
         $limit = $device->metadata[$this->deviceMetadataFieldLimit] ?? $this->deviceFallbackLimit;
         $diff = $this->timeDiffInSeconds($device->latestReceivedMessage->sentTime, $now);
         if ($diff >= $limit) {
-            $toMailAddress = $device->metadata[$this->deviceMetadataFieldMail] ?? ($application->contactEmail ?? $this->deviceFallbackMail);
-            $phone = $device->metadata[$this->deviceMetadataFieldPhone] ?? ($application->contactEmail ?? $this->deviceFallbackPhone);
-
             $subject = sprintf(
                 'Enhed "%s" offline siden %s',
                 $device->name,
@@ -127,7 +115,7 @@ final readonly class AlertManager
             );
             // Gateway limit for last seen is reached.
             $this->mailService->sendEmail(
-                to: $toMailAddress,
+                to: $this->findDeviceToMailAddress($device, $application, $overrideMail),
                 context: [
                     'device' => $device,
                     'since' => [
@@ -141,7 +129,18 @@ final readonly class AlertManager
                 textTemplate: 'device.txt.twig',
             );
 
-            // @todo: send sms
+            // SMS
+            $this->smsClient->send(
+                to: [$this->findDevicePhone($device, $application, $overridePhone)],
+                message: $this->templateService->renderTemplate('sms/gateway.twig', [
+                    'device' => $device,
+                    'since' => [
+                        'hours' => floor($diff / 3600),
+                        'minutes' => floor(($diff % 3600) / 60),
+                    ],
+                    'url' => sprintf($this->deviceBaseUrl, $device->applicationId, $device->id),
+                ])
+            );
         }
     }
 
@@ -170,9 +169,78 @@ final readonly class AlertManager
      * @return string
      *   The mail address
      */
-    private function findGatewayToMailAddress(Gateway $gateway): string
+    private function findGatewayToMailAddress(Gateway $gateway, string $overrideMail): string
     {
-        return $gateway->responsibleEmail ?? $this->gatewayFallbackMail;
+        if (!empty($overrideMail)) {
+            return $overrideMail;
+        }
+
+        return empty($gateway->responsibleEmail) ? $this->gatewayFallbackMail : $gateway->responsibleEmail;
+    }
+
+    /**
+     * Find the "to" e-mail-address with fallback address.
+     *
+     * @param Device $device
+     *   The device to look up metadata on
+     * @param Application|null $application
+     *   The application to fall back to
+     * @param string $overrideMail
+     *   The override mail address
+     *
+     * @return string
+     *   The mail address
+     */
+    private function findDeviceToMailAddress(Device $device, ?Application $application, string $overrideMail): string
+    {
+        if (!empty($overrideMail)) {
+            return $overrideMail;
+        }
+
+        return empty($device->metadata[$this->deviceMetadataFieldMail]) ? ($application->contactEmail ?? $this->deviceFallbackMail) : $device->metadata[$this->deviceMetadataFieldMail];
+    }
+
+    /**
+     * Find the phone number for a gateway.
+     *
+     * @param Gateway $gateway
+     *   The gateway to find number for
+     * @param string $overridePhone
+     *   An override value for the phone number
+     *
+     * @return string
+     *   The phone number to use, which is either the override phone, the gateway's
+     *   responsible phone, or a fallback device phone
+     */
+    public function findGatewayPhone(Gateway $gateway, string $overridePhone): string
+    {
+        if (!empty($overridePhone)) {
+            return $overridePhone;
+        }
+
+        return empty($gateway->responsiblePhone) ? $this->gatewayFallbackPhone : $gateway->responsiblePhone;
+    }
+
+    /**
+     * Find the phone number for a device.
+     *
+     * @param Device $device
+     *   Device to find number for
+     * @param ?Application $application
+     *   Application to fall back to
+     * @param string $overridePhone
+     *   Override number if given
+     *
+     * @return string
+     *   The number found
+     */
+    public function findDevicePhone(Device $device, ?Application $application, string $overridePhone): string
+    {
+        if (!empty($overridePhone)) {
+            return $overridePhone;
+        }
+
+        return empty($device->metadata[$this->deviceMetadataFieldPhone]) ? ($application->contactPhone ?? $this->deviceFallbackPhone) : $device->metadata[$this->deviceMetadataFieldPhone];
     }
 
     /**